requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
pandas
faker
plotly
tabulate